big pile of refactoring
